Load all of a mod's links
All four of a mod's links are now loaded, to coincide with the changes
in Paintjob Packer v1.7. Files are also now read using the new names,
including the mod authors' names. All vehicles return an empty trucky
link, which will be rectified once trucky is made public and its field
is added to every vehicle in Paintjob Packer.