Add pack editor tab
The description generation panel has been moved to a tab in a notebook,
and a new pack editing tab has been added. This will allow all of a
pack's variables to be updated as images are uploaded and as a mod is
uploaded to different websites, without needing a second window open
with the INI file.