Add checklist
A checklist has been added to the right side, with all of the steps
necessary for uploading a mod everywhere. A key has been added to each
pack's INI, representing how far along the process it is. Unsure if I'll
ever have to use it, as all of the steps should be completed at once,
but it's better to have it than to not.