Rewrite format_links
format_links() has been rewritten to work with descriptions containing
brackets and parentheses outside of a markdown link. Previously it
chopped them out as part of splitting the string into a list. At the
moment only the first link in a description will be formatted, and the
function will need to be tweaked to include a iterative loop if multiple
need to be supported in the future.