Add three buttons along the bottom
Three buttons have been added along the bottom of the window, accessible
regardless of which tab is open. The first opens the corresponding SCS
Forums thread for the currently selected game. The second saves the
pack, and the existing Save pack button has been removed, so that it is
accessible from both tabs. The third button clears the checklist to make
it faster to restart the uploading process when updating a pack.