Generate Ko-fi blog posts
A new "description" generator has been added to generate blog posts for
Ko-fi. Each post contains a list of packs updated since the last post
was made, and this list is collected as packs are updated (whenever the
Save pack button is clicked) in a new kofi-updates.txt, which has been
added to .gitignore. This list is emptied whenever a Ko-fi post is
generated, and its contents are printed to the console in case the
button was clicked by mistake, so it can be restored.